requirements.txt
asgiref==3.10.0
dj-database-url==2.3.0
Django==5.2.7
gunicorn==25.1.0
packaging==26.0
pillow==12.0.0
psycopg[binary]==3.2.10
qrcode==8.2
sqlparse==0.5.3
whitenoise==6.11.0
